# docs/swagger.yaml
definitions:
...
  api.UserSharePostRequest:
    properties:
      accessKey:
        type: string
      deviceID:
        type: string
      shareUserID:
        type: integer
    type: object
...
  util.CachePost:
    properties:
      data:
        type: string
      name:
        type: string
      sec:
        type: integer
    type: object
  util.ModelCaptcha:
    properties:
      captchaId:
        type: string
      captchaUrl:
        type: string
    type: object
info:
  contact: {}
paths:
...
  /api/user/share:
    get:
      consumes:
      - application/json
      description: 我的邀请列表或邀请的设备列表
      parameters:
      - in: query
        name: limit
        type: integer
      - in: query
        name: order
        type: string
      - in: query
        name: page
        type: integer
      - in: query
        name: user_id
        type: integer
      responses: {}
      summary: 老版本分享 已过期 不处理 不可删除
      tags:
      - 用户
    post:
      consumes:
      - application/json
      description: '- 提交成功邀请消息 邀请人，被邀请人UUID，数据签名。防止接口被发现刷接口。'
      parameters:
      - description: 参数列表
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/api.UserSharePostRequest'
      responses: {}
      security:
      - ApiKeyAuth: []
      summary: 提交成功邀请消息
      tags:
      - 用户
...
  /api/vip/list:
    get:
      consumes:
      - application/json
      description: 返回VIP列表数据
      parameters:
      - in: query
        name: limit
        type: integer
      - in: query
        name: onStatus
        type: boolean
      - in: query
        name: order
        type: string
      - in: query
        name: page
        type: integer
      responses: {}
      summary: VIP列表数据
      tags:
      - VIP
  /api/vip/unlock:
    post:
      consumes:
      - application/json
      description: 解锁一个新的视频VIP
      parameters:
      - description: 要解锁的VIPid
        in: query
        name: vipid
        required: true
        type: integer
      responses: {}
      summary: 解锁新的VIP
      tags:
      - VIP
...
  /util/cache:
    get:
      consumes:
      - application/json
      description: 获取redis的key
      parameters:
      - description: 查询IP地址
        in: query
        name: name
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/util.ModelCaptcha'
      summary: 获取数据
      tags:
      - 工具
    post:
      consumes:
      - application/json
      description: redis临时存储数据
      parameters:
      - description: 参数列表
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/util.CachePost'
      produces:
      - application/json
      responses: {}
      summary: 存储数据
      tags:
      - 工具
...
securityDefinitions:
  ApiKeyAuth:
    in: header
    name: x-token
    type: apiKey
swagger: "2.0"